dev_requirements.txt
pytest
numpy
requests
tqdm
python-dotenv
.
